+++ PWGEM/PhotonMeson/DataModel/gammaTables.h
