+++ PWGEM/PhotonMeson/TableProducer/skimmerGammaConversionsTruthOnlyMc.cxx
@@ -17,7 +17,7 @@
 // runme like: o2-analysis-trackselection -b --aod-file ${sourceFile} --aod-writer-json ${writerFile} | o2-analysis-timestamp -b | o2-analysis-trackextension -b | o2-analysis-lf-lambdakzerobuilder -b | o2-analysis-pid-tpc -b | o2-analysis-em-skimmermc -b
 
 // todo: remove reduantant information in GammaConversionsInfoTrue
-#include "gammaTables.h"
+#include "PWGEM/PhotonMeson/DataModel/gammaTables.h"
 
 #include "TVector3.h"
